Add files via upload
Most prominent addition to the program here is the uploading of the dataframe to a database in MySQL server.
I have also streamlined the spreadsheet and devised a plan to save the manipulated data to its own spreadsheet to avoid any confusion.
All blogs from the website have now been appended to the end of the spreadsheet. 
Some additional comments have been added to sections which are dealing with parsing data scraped from the blogs.