Add files via upload
I have ensured all of the SQL interactions are neater than before. This was achieved primarily by using a file with helper functions and call them when required to.
There is also a SQL query to obtain running totals for stakes, returns and profits then append them to another spreadsheet within the file. This allows us to track our work and results for the year 2023. 
I will be producing more graphs to outline company performance at a future date.